# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "CarKit API Documentation",
    "version": "0.0.1"
  },
  "host": [
    "carkit.lazyy.fr"
  ],
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/part/{carId}": {
      "post": {
        "tags": [
          "Part"
        ],
        "description": "Create a new custom part for a car.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Part creation data",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "Brakes"
                },
                "status": {
                  "example": "enable"
                },
                "advicedRevision": {
                  "example": "Check every 20,000 km"
                },
                "isCustom": {
                  "example": true
                }
              }
            }
          },
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "name": "Brakes",
              "status": "enable",
              "advicedRevision": "Check every 20,000 km",
              "isCustom": true
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "get": {
        "tags": [
          "Part"
        ],
        "description": "Get all parts for a car.",
        "parameters": [
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "name": "Brakes",
                "status": "enable",
                "advicedRevision": "Check every 20,000 km",
                "isCustom": true
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      }
    },
    "/part/{id}": {
      "put": {
        "tags": [
          "Part"
        ],
        "description": "Update a custom part.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Part update data",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "Updated Brakes"
                },
                "status": {
                  "example": "disabled"
                },
                "advicedRevision": {
                  "example": "Check every 15,000 km"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "name": "Updated Brakes",
              "status": "disabled",
              "advicedRevision": "Check every 15,000 km",
              "isCustom": true
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Part not found."
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Part"
        ],
        "description": "Delete a custom part.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Part deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Part not found."
            }
          }
        }
      }
    },
    "/mileage/{carId}": {
      "post": {
        "tags": [
          "Mileage"
        ],
        "description": "Add a new mileage entry for a car.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Mileage entry data",
            "schema": {
              "type": "object",
              "properties": {
                "mileage": {
                  "example": 15000
                },
                "date": {
                  "example": "2023-01-01"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "mileage": 15000,
              "date": "2023-01-01"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "get": {
        "tags": [
          "Mileage"
        ],
        "description": "Get all mileage entries for a car.",
        "parameters": [
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "mileage": 15000,
                "date": "2023-01-01"
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      }
    },
    "/mileage/{id}": {
      "delete": {
        "tags": [
          "Mileage"
        ],
        "description": "Delete a mileage entry.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Mileage entry deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Mileage entry not found."
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
}